requirements.txt
asgiref==3.3.1
Django==3.1.5
django-split-settings==1.0.1
pkg-resources==0.0.0
python-decouple==3.4
pytz==2020.5
sqlparse==0.4.1
